+++ packages/las-sdk-core/src/types/user.ts
@@ -1,12 +1,9 @@
 import { JSONValue, PaginationOptions, RequestConfig } from './common';
 
 export type User = {
-  avatar: string | null;
   createdBy: string | null;
   createdTime: string | null;
-  email: string;
   metadata: Record<string, JSONValue> | null;
-  name: string | null;
   profileId: string | null;
   roleIds: Array<string>;
   updatedBy: string | null;
@@ -16,16 +13,12 @@ export type User = {
 
 export type CreateUserOptions = RequestConfig & {
   appClientId?: string;
-  avatar?: string;
   metadata?: Record<string, JSONValue> | null;
-  name?: string;
   roleIds?: Array<string>;
 };
 
 export type UpdateUserOptions = RequestConfig & {
-  avatar?: string | null;
   metadata?: Record<string, JSONValue> | null;
-  name?: string | null;
   roleIds?: Array<string>;
 };